// frontend/src/types/enums.ts
export enum Gender {
    MALE = 'MALE',
    FEMALE = 'FEMALE',
    OTHER = 'OTHER'
}

export enum RelationshipType {
    OWNER = 'OWNER',
    SPOUSE = 'SPOUSE',
    CHILD = 'CHILD',
    PARENT = 'PARENT',
    SIBLING = 'SIBLING',
    OTHER = 'OTHER'
}

export enum ChangeType {
    TEMPORARY_ABSENCE = 'TEMPORARY_ABSENCE',
    PERMANENT_ABSENCE = 'PERMANENT_ABSENCE',
    NEW_RESIDENT = 'NEW_RESIDENT',
    RETURN = 'RETURN'
} 